requirements.txt
fastapi
tortoise-orm
uvicorn[standard]
tortoise-orm[aiomysql]
requests
